Disable automatic email after the gift purchase
The default is not very straightforward and requires integration
with Mailer and specific mail template. For that reason we made
it disabled by default with the option to enable it back.

To complement the feature by ScenariosModule, we're adding new
condition to differentiate payments based on their items.

remp/crm#2199