Conditional DI registration of scenarios donor notification
The scenarios notification is directly dependent on REMP mailer
module, because it needs to display list of available emails.
However gifts module shoudln't have direct dependency
on REMP mailer and it would crash existing installations.

This commit moves the DI registration of generic event,
and its registration to the scenario generics manager
to GiftsModuleExtension, which executes it only when class
from REMP Mailer module is actually available.